feat(shared): move props logic to extensuble-entity
