+++ packages/shared/src/index.ts
@@ -1,3 +1,2 @@
-export type { ExtensiblePropDescription } from './props';
-export { ExtensibleEntity } from './extensible-entity';
+export * from './extensible-entity';
 export * from './error';
